# web_server/docs/swagger.yaml
definitions:
...
  response.Body:
    properties:
      code:
        type: integer
      data: {}
      msg:
        type: string
      status:
        type: boolean
    type: object
info:
  contact: {}
  description: 基于 Vue3+UniApp+Gin+Gorm 的社团管理系统后端接口
  title: 高校社团信息管理系统 API
  version: "1.0"
paths:
...
  /leader/attendance/{id}:
    delete:
      parameters:
      - description: 考勤记录ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Body'
      security:
      - Bearer: []
      summary: 删除异常考勤记录（负责人）
      tags:
      - 考勤
  /leader/clubs/{clubId}/attendance:
    get:
      parameters:
      - description: 社团ID
        in: path
        name: clubId
        required: true
        type: integer
      - description: 用户ID
        in: query
        name: userId
        type: integer
      - description: 页码
        in: query
        name: page
        type: integer
      - description: 每页数量
        in: query
        name: pageSize
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Body'
      security:
      - Bearer: []
      summary: 社团考勤检索（负责人）
      tags:
      - 考勤
...
  /leader/clubs/{clubId}/members/users:
    get:
      parameters:
      - description: 社团ID
        in: path
        name: clubId
        required: true
        type: integer
      - description: '角色: member/leader/advisor'
        in: query
        name: role
        type: string
      - description: '状态: pending/approved/rejected/quit'
        in: query
        name: status
        type: string
      - description: 关键词：姓名/学号/手机号
        in: query
        name: keyword
        type: string
      - description: 页码
        in: query
        name: page
        type: integer
      - description: 每页数量
        in: query
        name: pageSize
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Body'
      security:
      - Bearer: []
      summary: 社团成员列表（负责人）
      tags:
      - 成员
  /leader/clubs/{clubId}/memberships:
    get:
      parameters:
      - description: 社团ID
        in: path
        name: clubId
        required: true
        type: integer
      - description: 页码
        in: query
        name: page
        type: integer
      - description: 每页数量
        in: query
        name: pageSize
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Body'
      security:
      - Bearer: []
      summary: 待审批入会列表（负责人）
      tags:
      - 成员
  /leader/clubs/{clubId}/memberships/{id}/approve:
    post:
      consumes:
      - application/json
      parameters:
      - description: 社团ID
        in: path
        name: clubId
        required: true
        type: integer
      - description: 成员关系ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Body'
      security:
      - Bearer: []
      summary: 审批通过成员（负责人）
      tags:
      - 成员
  /leader/clubs/{clubId}/memberships/{id}/reject:
    post:
      consumes:
      - application/json
      parameters:
      - description: 社团ID
        in: path
        name: clubId
        required: true
        type: integer
      - description: 成员关系ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Body'
      security:
      - Bearer: []
      summary: 审批驳回成员（负责人）
      tags:
      - 成员
...
  /member/activities/{activityId}/signin:
    post:
      parameters:
      - description: 活动ID
        in: path
        name: activityId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Body'
      security:
      - Bearer: []
      summary: 会员签到
      tags:
      - 考勤
  /member/activities/{activityId}/signout:
    post:
      parameters:
      - description: 活动ID
        in: path
        name: activityId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Body'
      security:
      - Bearer: []
      summary: 会员签退
      tags:
      - 考勤
  /member/attendance/my:
    get:
      parameters:
      - description: 社团ID
        in: query
        name: clubId
        required: true
        type: integer
      - description: 页码
        in: query
        name: page
        type: integer
      - description: 每页数量
        in: query
        name: pageSize
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Body'
      security:
      - Bearer: []
      summary: 我的考勤记录（按社团）
      tags:
      - 考勤
...
  /public/clubs:
    get:
      parameters:
      - description: 页码
        in: query
        name: page
        type: integer
      - description: 每页数量
        in: query
        name: pageSize
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Body'
      summary: 社团列表（含类别、代表活动、负责人联系方式）
      tags:
      - 公共
...
  /student/clubs/{clubId}/apply:
    post:
      parameters:
      - description: 社团ID
        in: path
        name: clubId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Body'
      security:
      - Bearer: []
      summary: 申请加入社团
      tags:
      - 学生
  /student/clubs/{clubId}/exit:
    post:
      parameters:
      - description: 社团ID
        in: path
        name: clubId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Body'
      security:
      - Bearer: []
      summary: 退出社团
      tags:
      - 学生
  /student/memberships/my:
    get:
      parameters:
      - description: '状态: pending/approved/rejected/quit'
        in: query
        name: status
        type: string
      - description: 页码
        in: query
        name: page
        type: integer
      - description: 每页数量
        in: query
        name: pageSize
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Body'
      security:
      - Bearer: []
      summary: 我的社团关系
      tags:
      - 学生
...
securityDefinitions:
  Bearer:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"